Issue #23: Remove all testing related code and tasks from this repository
