--- .github/workflows/test-static-tests.yml
@@ -1,38 +1,0 @@
-name: Test Static Tests
-on:
-  push:
-    branches:
-      - main
-  pull_request:
-    types: [opened, synchronize, reopened]
-
-jobs:
-  Test-Static-Tests:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v2
-      - uses: shivammathur/setup-php@v2
-        with:
-          php-version: '7.4'
-          tools: composer:v2
-          extensions: gd
-      - name: Get composer cache directory
-        id: composer-cache
-        run: echo "::set-output name=dir::$(composer config cache-files-dir)"
-      - name: Cache dependencies
-        uses: actions/cache@v2
-        with:
-          path: ${{ steps.composer-cache.outputs.dir }}
-          key: ${{ runner.os }}-composer-${{ hashFiles('**/composer.lock') }}
-          restore-keys: ${{ runner.os }}-composer-
-      - name: Install Drupal
-        run: |
-          cd ../
-          mkdir drupal
-          cd drupal
-          cp ${GITHUB_WORKSPACE}/tests/fixtures.drainpipe-external-test-project/composer.json .
-          composer install
-      - name: Run static tests
-        run: |
-          cd ../drupal
-          ./vendor/bin/task test:static
