--- config/phpstan.neon
@@ -1,11 +1,0 @@
-includes:
-	- ../../../mglaman/phpstan-drupal/extension.neon
-
-parameters:
-	# Ignore tests
-	excludes_analyse:
-		- *Test.php
-		- *TestBase.php
-		- *default.settings.php
-	# PHPStan Level 1
-	level: 1
