+++ tests/fixtures.drainpipe-external-test-project/composer.json
@@ -26,9 +26,7 @@
     "drupal/core": "^9.2",
     "drupal/core-composer-scaffold": "^9.2",
     "drupal/core-project-message": "^9.2",
-    "drupal/core-recommended": "^9.2"
-  },
-  "require-dev": {
+    "drupal/core-recommended": "^9.2",
     "lullabot/drainpipe": "*"
   },
   "conflict": {
